5. SAP Generative AI Hub and HANA Cloud Vector Engine/5.1 Q&A RAG/code/qna-app/requirements.txt
pandas==2.2.2
pydantic==2.9.1
pydantic_core==2.23.3
requests==2.32.3
streamlit==1.38.0
urllib3==2.2.2
python-dotenv
ai-core-sdk
openai